requirements.txt
flask==2.0.1
Jinja2==2.11.2
Werkzeug==1.0.1
PyJWT==2.3.0
requests==2.28.2
psycopg2-binary==2.9.1